backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
python-dotenv==1.0.1
openai==1.35.0
PyJWT==2.8.0
beautifulsoup4==4.12.3
httpx==0.27.2
psycopg[binary]==3.2.9
pgvector==0.3.3
